renamed package path, add _beta tags, added some comment on autoMl model
