+++ samples/snippets/src/main/java/documentai/v1beta2/BatchParseTableBeta.java
@@ -14,24 +14,26 @@
  * limitations under the License.
  */
 
-package com.examples.documentai;
+package documentai.v1beta2;
 
-// [START documentai_batch_parse_form]
+// [START documentai_batch_parse_table_beta]
 
 import com.google.api.gax.longrunning.OperationFuture;
 import com.google.api.gax.paging.Page;
 import com.google.cloud.documentai.v1beta2.BatchProcessDocumentsRequest;
 import com.google.cloud.documentai.v1beta2.BatchProcessDocumentsResponse;
+import com.google.cloud.documentai.v1beta2.BoundingPoly;
 import com.google.cloud.documentai.v1beta2.Document;
 import com.google.cloud.documentai.v1beta2.DocumentUnderstandingServiceClient;
-import com.google.cloud.documentai.v1beta2.FormExtractionParams;
 import com.google.cloud.documentai.v1beta2.GcsDestination;
 import com.google.cloud.documentai.v1beta2.GcsSource;
 import com.google.cloud.documentai.v1beta2.InputConfig;
-import com.google.cloud.documentai.v1beta2.KeyValuePairHint;
+import com.google.cloud.documentai.v1beta2.NormalizedVertex;
 import com.google.cloud.documentai.v1beta2.OperationMetadata;
 import com.google.cloud.documentai.v1beta2.OutputConfig;
 import com.google.cloud.documentai.v1beta2.ProcessDocumentRequest;
+import com.google.cloud.documentai.v1beta2.TableBoundHint;
+import com.google.cloud.documentai.v1beta2.TableExtractionParams;
 import com.google.cloud.storage.Blob;
 import com.google.cloud.storage.BlobId;
 import com.google.cloud.storage.Bucket;
@@ -41,13 +43,14 @@ import com.google.protobuf.util.JsonFormat;
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
+import java.util.List;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
-public class BatchParseForm {
+public class BatchParseTableBeta {
 
-  public static void batchParseFormGcs()
+  public static void batchParseTableGcs()
       throws IOException, InterruptedException, ExecutionException, TimeoutException {
     // TODO(developer): Replace these variables before running the sample.
     String projectId = "your-project-id";
@@ -55,10 +58,10 @@ public class BatchParseForm {
     String outputGcsBucketName = "your-gcs-bucket-name";
     String outputGcsPrefix = "PREFIX";
     String inputGcsUri = "gs://your-gcs-bucket/path/to/input/file.json";
-    batchParseFormGcs(projectId, location, outputGcsBucketName, outputGcsPrefix, inputGcsUri);
+    batchParseTableGcs(projectId, location, outputGcsBucketName, outputGcsPrefix, inputGcsUri);
   }
 
-  public static void batchParseFormGcs(
+  public static void batchParseTableGcs(
       String projectId,
       String location,
       String outputGcsBucketName,
@@ -68,34 +71,33 @@ public class BatchParseForm {
     // Initialize client that will be used to send requests. This client only needs to be created
     // once, and can be reused for multiple requests. After completing all of your requests, call
     // the "close" method on the client to safely clean up any remaining background resources.
-    try (DocumentUnderstandingServiceClient client =
-        DocumentUnderstandingServiceClient.create()) {
+    try (DocumentUnderstandingServiceClient client = DocumentUnderstandingServiceClient.create()) {
 
       // Configure the request for processing the PDF
       String parent = String.format("projects/%s/locations/%s", projectId, location);
 
-      // Improve form parsing results by providing key-value pair hints.
-      // For each key hint, key is text that is likely to appear in the
-      // document as a form field name (i.e. "DOB").
-      // Value types are optional, but can be one or more of:
-      // ADDRESS, LOCATION, ORGANIZATION, PERSON, PHONE_NUMBER, ID,
-      // NUMBER, EMAIL, PRICE, TERMS, DATE, NAME
-      KeyValuePairHint keyValuePairHint =
-          KeyValuePairHint.newBuilder().setKey("Phone").addValueTypes("PHONE_NUMBER").build();
-
-      KeyValuePairHint keyValuePairHint2 =
-          KeyValuePairHint.newBuilder()
-              .setKey("Contact")
-              .addValueTypes("EMAIL")
-              .addValueTypes("NAME")
+      TableBoundHint tableBoundHints =
+          TableBoundHint.newBuilder()
+              .setBoundingBox(
+                  // Define a polygon around tables to detect
+                  // Each vertice coordinate must be a number between 0 and 1
+                  BoundingPoly.newBuilder()
+                      // top left
+                      .addNormalizedVertices(NormalizedVertex.newBuilder().setX(0).setX(0).build())
+                      // top right
+                      .addNormalizedVertices(NormalizedVertex.newBuilder().setX(1).setX(0).build())
+                      // bottom right
+                      .addNormalizedVertices(NormalizedVertex.newBuilder().setX(1).setX(1).build())
+                      // bottom left
+                      .addNormalizedVertices(NormalizedVertex.newBuilder().setX(0).setX(1).build())
+                      .build())
+              .setPageNumber(1)
               .build();
 
-      // Setting enabled=True enables form extraction
-      FormExtractionParams params =
-          FormExtractionParams.newBuilder()
+      TableExtractionParams params =
+          TableExtractionParams.newBuilder()
               .setEnabled(true)
-              .addKeyValuePairHints(keyValuePairHint)
-              .addKeyValuePairHints(keyValuePairHint2)
+              .addTableBoundHints(tableBoundHints)
               .build();
 
       GcsSource inputUri = GcsSource.newBuilder().setUri(inputGcsUri).build();
@@ -103,20 +105,19 @@ public class BatchParseForm {
       // mime_type can be application/pdf, image/tiff,
       // and image/gif, or application/json
       InputConfig config =
-          InputConfig.newBuilder().setGcsSource(inputUri)
-                  .setMimeType("application/pdf").build();
+          InputConfig.newBuilder().setGcsSource(inputUri).setMimeType("application/pdf").build();
 
-      GcsDestination gcsDestination = GcsDestination.newBuilder()
-              .setUri(String.format("gs://%s/%s", outputGcsBucketName, outputGcsPrefix)).build();
-
-      OutputConfig outputConfig =  OutputConfig.newBuilder()
-              .setGcsDestination(gcsDestination)
-              .setPagesPerShard(1)
+      GcsDestination gcsDestination =
+          GcsDestination.newBuilder()
+              .setUri(String.format("gs://%s/%s", outputGcsBucketName, outputGcsPrefix))
               .build();
+
+      OutputConfig outputConfig =
+          OutputConfig.newBuilder().setGcsDestination(gcsDestination).setPagesPerShard(1).build();
 
       ProcessDocumentRequest request =
           ProcessDocumentRequest.newBuilder()
-              .setFormExtractionParams(params)
+              .setTableExtractionParams(params)
               .setInputConfig(config)
               .setOutputConfig(outputConfig)
               .build();
@@ -158,7 +159,6 @@ public class BatchParseForm {
           FileReader reader = new FileReader(tempFile);
           Document.Builder builder = Document.newBuilder();
           JsonFormat.parser().merge(reader, builder);
-
           Document document = builder.build();
 
           // Get all of the document text as one big string.
@@ -166,12 +166,26 @@ public class BatchParseForm {
 
           // Process the output.
           Document.Page page1 = document.getPages(0);
-          for (Document.Page.FormField field : page1.getFormFieldsList()) {
-            String fieldName = getText(field.getFieldName(), text);
-            String fieldValue = getText(field.getFieldValue(), text);
+          Document.Page.Table table = page1.getTables(0);
 
-            System.out.println("Extracted form fields pair:");
-            System.out.printf("\t(%s, %s))", fieldName, fieldValue);
+          System.out.println("Results from first table processed:");
+          System.out.println(
+              "First detected language: " + page1.getDetectedLanguages(0).getLanguageCode());
+          System.out.println("Header row:");
+
+          Document.Page.Table.TableRow headerRow = table.getHeaderRows(0);
+
+          for (Document.Page.Table.TableCell tableCell : headerRow.getCellsList()) {
+            if (!tableCell.getLayout().getTextAnchor().getTextSegmentsList().isEmpty()) {
+              // Extract shards from the text field
+              // First shard in document doesn't have startIndex property
+              List<Document.TextAnchor.TextSegment> textSegments =
+                  tableCell.getLayout().getTextAnchor().getTextSegmentsList();
+              int startIdx =
+                  textSegments.size() > 0 ? (int) textSegments.get(0).getStartIndex() : 0;
+              int endIdx = (int) textSegments.get(0).getEndIndex();
+              System.out.printf("\t%s", text.substring(startIdx, endIdx));
+            }
           }
 
           // Clean up temp file.
@@ -180,15 +194,5 @@ public class BatchParseForm {
       }
     }
   }
-
-  private static String getText(Document.Page.Layout layout, String text) {
-    Document.TextAnchor textAnchor = layout.getTextAnchor();
-    if (textAnchor.getTextSegmentsList().size() > 0) {
-      int startIdx = (int) textAnchor.getTextSegments(0).getStartIndex();
-      int endIdx = (int) textAnchor.getTextSegments(0).getEndIndex();
-      return text.substring(startIdx, endIdx);
-    }
-    return "[NO TEXT]";
-  }
 }
-// [END documentai_batch_parse_form]
+// [END documentai_batch_parse_table_beta]
