+++ samples/snippets/src/main/java/documentai/v1beta2/ParseTableBeta.java
@@ -14,31 +14,34 @@
  * limitations under the License.
  */
 
-package com.examples.documentai;
+package documentai.v1beta2;
 
-// [START documentai_parse_form]
+// [START documentai_parse_table_beta]
 
+import com.google.cloud.documentai.v1beta2.BoundingPoly;
 import com.google.cloud.documentai.v1beta2.Document;
 import com.google.cloud.documentai.v1beta2.DocumentUnderstandingServiceClient;
-import com.google.cloud.documentai.v1beta2.FormExtractionParams;
 import com.google.cloud.documentai.v1beta2.GcsSource;
 import com.google.cloud.documentai.v1beta2.InputConfig;
-import com.google.cloud.documentai.v1beta2.KeyValuePairHint;
+import com.google.cloud.documentai.v1beta2.NormalizedVertex;
 import com.google.cloud.documentai.v1beta2.ProcessDocumentRequest;
+import com.google.cloud.documentai.v1beta2.TableBoundHint;
+import com.google.cloud.documentai.v1beta2.TableExtractionParams;
 import java.io.IOException;
-import java.util.concurrent.ExecutionException;
+import java.util.List;
 
-public class ParseForm {
-  public static void parseForm() throws IOException, ExecutionException, InterruptedException {
+public class ParseTableBeta {
+
+  public static void parseTable() throws IOException {
     // TODO(developer): Replace these variables before running the sample.
     String projectId = "your-project-id";
     String location = "your-project-location"; // Format is "us" or "eu".
     String inputGcsUri = "gs://your-gcs-bucket/path/to/input/file.json";
-    parseForm(projectId, location, inputGcsUri);
+    parseTable(projectId, location, inputGcsUri);
   }
 
-  public static void parseForm(String projectId, String location, String inputGcsUri)
-      throws IOException, ExecutionException, InterruptedException {
+  public static void parseTable(String projectId, String location, String inputGcsUri)
+      throws IOException {
     // Initialize client that will be used to send requests. This client only needs to be created
     // once, and can be reused for multiple requests. After completing all of your requests, call
     // the "close" method on the client to safely clean up any remaining background resources.
@@ -46,27 +49,28 @@ public class ParseForm {
       // Configure the request for processing the PDF
       String parent = String.format("projects/%s/locations/%s", projectId, location);
 
-      // Improve form parsing results by providing key-value pair hints.
-      // For each key hint, key is text that is likely to appear in the
-      // document as a form field name (i.e. "DOB").
-      // Value types are optional, but can be one or more of:
-      // ADDRESS, LOCATION, ORGANIZATION, PERSON, PHONE_NUMBER, ID,
-      // NUMBER, EMAIL, PRICE, TERMS, DATE, NAME
-      KeyValuePairHint keyValuePairHint =
-          KeyValuePairHint.newBuilder().setKey("Phone").addValueTypes("PHONE_NUMBER").build();
-      KeyValuePairHint keyValuePairHint2 =
-          KeyValuePairHint.newBuilder()
-              .setKey("Contact")
-              .addValueTypes("EMAIL")
-              .addValueTypes("NAME")
+      TableBoundHint tableBoundHints =
+          TableBoundHint.newBuilder()
+              .setBoundingBox(
+                  // Define a polygon around tables to detect
+                  // Each vertice coordinate must be a number between 0 and 1
+                  BoundingPoly.newBuilder()
+                      // top left
+                      .addNormalizedVertices(NormalizedVertex.newBuilder().setX(0).setX(0).build())
+                      // top right
+                      .addNormalizedVertices(NormalizedVertex.newBuilder().setX(1).setX(0).build())
+                      // bottom right
+                      .addNormalizedVertices(NormalizedVertex.newBuilder().setX(1).setX(1).build())
+                      // bottom left
+                      .addNormalizedVertices(NormalizedVertex.newBuilder().setX(0).setX(1).build())
+                      .build())
+              .setPageNumber(1)
               .build();
 
-      // Setting enabled=True enables form extraction
-      FormExtractionParams params =
-          FormExtractionParams.newBuilder()
+      TableExtractionParams params =
+          TableExtractionParams.newBuilder()
               .setEnabled(true)
-              .addKeyValuePairHints(keyValuePairHint)
-              .addKeyValuePairHints(keyValuePairHint2)
+              .addTableBoundHints(tableBoundHints)
               .build();
 
       GcsSource uri = GcsSource.newBuilder().setUri(inputGcsUri).build();
@@ -79,7 +83,7 @@ public class ParseForm {
       ProcessDocumentRequest request =
           ProcessDocumentRequest.newBuilder()
               .setParent(parent)
-              .setFormExtractionParams(params)
+              .setTableExtractionParams(params)
               .setInputConfig(config)
               .build();
 
@@ -89,18 +93,30 @@ public class ParseForm {
       // Get all of the document text as one big string
       String text = response.getText();
 
-      // Process the output
+      // Get the first table in the document
       Document.Page page1 = response.getPages(0);
-      for (Document.Page.FormField field : page1.getFormFieldsList()) {
-        String fieldName = getText(field.getFieldName(), text);
-        String fieldValue = getText(field.getFieldValue(), text);
+      Document.Page.Table table = page1.getTables(0);
 
-        System.out.println("Extracted form fields pair:");
-        System.out.printf("\t(%s, %s))", fieldName, fieldValue);
+      System.out.println("Results from first table processed:");
+      List<Document.Page.DetectedLanguage> detectedLangs = page1.getDetectedLanguagesList();
+      String langCode =
+          detectedLangs.size() > 0 ? detectedLangs.get(0).getLanguageCode() : "NOT_FOUND";
+      System.out.printf("First detected language: : %s", langCode);
+
+      Document.Page.Table.TableRow headerRow = table.getHeaderRows(0);
+      System.out.println("Header row:");
+
+      for (Document.Page.Table.TableCell tableCell : headerRow.getCellsList()) {
+        if (tableCell.getLayout().getTextAnchor().getTextSegmentsList() != null) {
+          // Extract shards from the text field
+          // First shard in document doesn't have startIndex property
+          System.out.printf("\t%s", getText(tableCell.getLayout(), text));
+        }
       }
     }
   }
 
+  // Extract shards from the text field
   private static String getText(Document.Page.Layout layout, String text) {
     Document.TextAnchor textAnchor = layout.getTextAnchor();
     if (textAnchor.getTextSegmentsList().size() > 0) {
@@ -111,4 +127,4 @@ public class ParseForm {
     return "[NO TEXT]";
   }
 }
-// [END documentai_parse_form]
+// [END documentai_parse_table_beta]
