+++ samples/snippets/src/main/java/documentai/v1beta2/ParseWithModelBeta.java
@@ -14,10 +14,11 @@
  * limitations under the License.
  */
 
-package com.examples.documentai;
+package documentai.v1beta2;
 
-// [START documentai_quickstart]
+// [START documentai_parse_with_model_beta]
 
+import com.google.cloud.documentai.v1beta2.AutoMlParams;
 import com.google.cloud.documentai.v1beta2.Document;
 import com.google.cloud.documentai.v1beta2.DocumentUnderstandingServiceClient;
 import com.google.cloud.documentai.v1beta2.GcsSource;
@@ -25,26 +26,30 @@ import com.google.cloud.documentai.v1beta2.InputConfig;
 import com.google.cloud.documentai.v1beta2.ProcessDocumentRequest;
 import java.io.IOException;
 
-public class QuickStart {
+public class ParseWithModelBeta {
 
-  public static void quickStart() throws IOException {
+  public static void parseWithModel() throws IOException {
     // TODO(developer): Replace these variables before running the sample.
     String projectId = "your-project-id";
     String location = "your-project-location"; // Format is "us" or "eu".
-    String inputGcsUri = "gs://your-gcs-bucket/path/to/input/file.json";
-    quickStart(projectId, location, inputGcsUri);
+    // AutoML model name formatted as: "projects/[PROJECT_ID]/locations/[LOCATION]/models/[MODEL_ID]"
+    String autoMlModel = "your-full-resource-model-name";
+    String gcsUri = "gs://your-gcs-bucket/path/to/input/file.json";
+    parseWithModel(projectId, location, autoMlModel, gcsUri);
   }
 
-  public static void quickStart(String projectId, String location, String inputGcsUri)
-      throws IOException {
+  public static void parseWithModel(
+      String projectId, String location, String autoMlModel, String gcsUri) throws IOException {
     // Initialize client that will be used to send requests. This client only needs to be created
     // once, and can be reused for multiple requests. After completing all of your requests, call
     // the "close" method on the client to safely clean up any remaining background resources.
     try (DocumentUnderstandingServiceClient client = DocumentUnderstandingServiceClient.create()) {
-      // Configure the request for processing a single document
+      // Configure the request for processing the PDF
       String parent = String.format("projects/%s/locations/%s", projectId, location);
 
-      GcsSource uri = GcsSource.newBuilder().setUri(inputGcsUri).build();
+      AutoMlParams params = AutoMlParams.newBuilder().setModel(autoMlModel).build();
+
+      GcsSource uri = GcsSource.newBuilder().setUri(gcsUri).build();
 
       // mime_type can be application/pdf, image/tiff,
       // and image/gif, or application/json
@@ -52,27 +57,21 @@ public class QuickStart {
           InputConfig.newBuilder().setGcsSource(uri).setMimeType("application/pdf").build();
 
       ProcessDocumentRequest request =
-          ProcessDocumentRequest.newBuilder().setParent(parent).setInputConfig(config).build();
+          ProcessDocumentRequest.newBuilder()
+              .setParent(parent)
+              .setAutomlParams(params)
+              .setInputConfig(config)
+              .build();
 
       // Recognizes text entities in the PDF document
       Document response = client.processDocument(request);
 
-      // Get all of the document text as one big string
-      String text = response.getText();
-
       // Process the output
-      for (Document.Entity entity : response.getEntitiesList()) {
-        System.out.printf("Entity text: %s\n", getText(entity, text));
-        System.out.printf("Entity type: %s\n", entity.getType());
-        System.out.printf("Entity mention text: %s\n", entity.getMentionText());
+      for (Document.Label label : response.getLabelsList()) {
+        System.out.printf("Label detected: %s\n", label.getName());
+        System.out.printf("Confidence:  %s\n", label.getConfidence());
       }
     }
   }
-
-  private static String getText(Document.Entity entity, String text) {
-    int startIdx = (int) entity.getTextAnchor().getTextSegments(0).getStartIndex();
-    int endIdx = (int) entity.getTextAnchor().getTextSegments(0).getEndIndex();
-    return text.substring(startIdx, endIdx);
-  }
 }
-// [END documentai_quickstart]
+// [END documentai_parse_with_model_beta]
