+++ samples/snippets/src/test/java/documentai/v1beta2/ParseTableBetaTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.examples.documentai;
+package documentai.v1beta2;
 
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.Assert.assertNotNull;
@@ -26,7 +26,7 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
-public class ParseTableTest {
+public class ParseTableBetaTest {
   private static final String PROJECT_ID = System.getenv("GCLOUD_PROJECT");
   private static final String INPUT_URI = "gs://cloud-samples-data/documentai/invoice.pdf";
 
@@ -55,7 +55,7 @@ public class ParseTableTest {
   @Test
   public void testParseTable() throws IOException {
     // parse the GCS invoice as a table.
-    ParseTable.parseTable(PROJECT_ID,"us", INPUT_URI);
+    ParseTableBeta.parseTable(PROJECT_ID,"us", INPUT_URI);
     String got = bout.toString();
 
     assertThat(got).contains("First detected language");
